# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkHighlightReviewRequest": {
        "properties": {
          "highlight_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Highlight Ids",
            "type": "array"
          },
          "status": {
            "enum": [
              "accepted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "highlight_ids",
          "status"
        ],
        "title": "BulkHighlightReviewRequest",
        "type": "object"
      },
...
      "CommentReactionRequest": {
        "properties": {
          "reaction": {
            "enum": [
              "acknowledged",
              "agree",
              "question"
            ],
            "title": "Reaction",
            "type": "string"
          }
        },
        "required": [
          "reaction"
        ],
        "title": "CommentReactionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HighlightResponse": {
        "properties": {
          "category": {
            "default": "risk",
            "enum": [
              "risk",
              "symptom",
              "medication",
              "care_gap",
              "patient_context",
              "follow_up"
            ],
            "title": "Category",
            "type": "string"
          },
          "clinic_id": {
            "format": "uuid",
            "title": "Clinic Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duplicate_group_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Group Id"
          },
          "generated_by": {
            "enum": [
              "rule",
              "ai",
              "clinician"
            ],
            "title": "Generated By",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "normalized_claim": {
            "title": "Normalized Claim",
            "type": "string"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "quoted_text": {
            "title": "Quoted Text",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "risk_level": {
            "enum": [
              "information",
              "attention",
              "critical"
            ],
            "title": "Risk Level",
            "type": "string"
          },
          "risk_reason": {
            "title": "Risk Reason",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_end_offset": {
            "title": "Source End Offset",
            "type": "integer"
          },
          "source_entry_id": {
            "format": "uuid",
            "title": "Source Entry Id",
            "type": "string"
          },
          "source_start_offset": {
            "title": "Source Start Offset",
            "type": "integer"
          },
          "source_version_id": {
            "format": "uuid",
            "title": "Source Version Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "suggested",
              "accepted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clinic_id",
          "patient_id",
          "source_entry_id",
          "source_version_id",
          "source_start_offset",
          "source_end_offset",
          "quoted_text",
          "normalized_claim",
          "risk_level",
          "risk_reason",
          "score",
          "status",
          "generated_by",
          "reviewed_by",
          "reviewed_at",
          "created_at"
        ],
        "title": "HighlightResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Server-enforced API for the Nightingale longitudinal Care Note.",
    "title": "Nightingale Care Note API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/comments/{comment_id}/reactions": {
      "post": {
        "operationId": "add_comment_reaction_comments__comment_id__reactions_post",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Comment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Comment Reaction Comments  Comment Id  Reactions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Comment Reaction",
        "tags": [
          "collaboration"
        ]
      }
    },
    "/comments/{comment_id}/reactions/{reaction}": {
      "delete": {
        "operationId": "remove_comment_reaction_comments__comment_id__reactions__reaction__delete",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Comment Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "reaction",
            "required": true,
            "schema": {
              "title": "Reaction",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Remove Comment Reaction Comments  Comment Id  Reactions  Reaction  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Comment Reaction",
        "tags": [
          "collaboration"
        ]
      }
    },
...
    "/highlights/bulk-review": {
      "post": {
        "operationId": "bulk_review_highlights_highlights_bulk_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkHighlightReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HighlightResponse"
                  },
                  "title": "Response Bulk Review Highlights Highlights Bulk Review Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bulk Review Highlights",
        "tags": [
          "highlights"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "list_notifications_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "title": "Response List Notifications Notifications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{notification_id}/dismiss": {
      "post": {
        "operationId": "dismiss_notification_notifications__notification_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dismiss Notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/patient-summary-reviews/{review_id}/review": {
      "post": {
        "operationId": "review_patient_summary_patient_summary_reviews__review_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientSummaryReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSummaryReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Patient Summary",
        "tags": [
          "AI scribe review"
        ]
      }
    },
...
    "/patients/{patient_id}/audit.csv": {
      "get": {
        "operationId": "download_audit_report_patients__patient_id__audit_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Audit Report",
        "tags": [
          "audit"
        ]
      }
    },
    "/patients/{patient_id}/comments": {
      "get": {
        "operationId": "list_comments_patients__patient_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommentResponse"
                  },
                  "title": "Response List Comments Patients  Patient Id  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Comments",
        "tags": [
          "collaboration"
        ]
      },
      "post": {
        "operationId": "create_comment_patients__patient_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Comment",
        "tags": [
          "collaboration"
        ]
      }
    },
...
    "/patients/{patient_id}/highlight-groups": {
      "get": {
        "operationId": "duplicate_highlight_groups_patients__patient_id__highlight_groups_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Duplicate Highlight Groups Patients  Patient Id  Highlight Groups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Duplicate Highlight Groups",
        "tags": [
          "highlights"
        ]
      }
    },
    "/patients/{patient_id}/highlights": {
      "get": {
        "operationId": "list_highlights_patients__patient_id__highlights_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HighlightResponse"
                  },
                  "title": "Response List Highlights Patients  Patient Id  Highlights Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Highlights",
        "tags": [
          "highlights"
        ]
      }
    },
    "/patients/{patient_id}/patient-summary-reviews": {
      "get": {
        "operationId": "list_patient_summary_reviews_patients__patient_id__patient_summary_reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PatientSummaryReviewResponse"
                  },
                  "title": "Response List Patient Summary Reviews Patients  Patient Id  Patient Summary Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patient Summary Reviews",
        "tags": [
          "AI scribe review"
        ]
      }
    },
...
  }
}